Created abstract validators for min and max lengths and created package for all validators. Resolves GH-80
