+++ data/validators/IsClassIdExisting.ts
@@ -5,18 +5,18 @@ import {
     ValidatorConstraint,
     ValidatorConstraintInterface
 } from "class-validator";
-import { MysqlDatabase } from "../../../database/MysqlDatabase";
-import { ReportsModel } from "../../../models/ReportsModel";
+import { MysqlDatabase } from "../../database/MysqlDatabase";
+import { ClassesModel } from "../../models/ClassesModel";
 
 @ValidatorConstraint({async: true})
-export class IsIdExistingConstraint implements ValidatorConstraintInterface {
+export class IsClassIdExistingConstraint implements ValidatorConstraintInterface {
     validate(id: number, validationArguments?: ValidationArguments): Promise<boolean> | boolean {
         return new Promise(async resolve => {
             if (!id) {
                 resolve(true);
             }
 
-            const result = await new ReportsModel(new MysqlDatabase())
+            const result = await new ClassesModel(new MysqlDatabase())
                 .findById(id);
 
             resolve(!!result);
@@ -24,14 +24,14 @@ export class IsIdExistingConstraint implements ValidatorConstraintInterface {
     }
 }
 
-export function IsIdExisting(validationOptions: ValidationOptions) {
+export function IsClassIdExisting(validationOptions: ValidationOptions) {
     return function (object: Object, propertyName: string) {
         registerDecorator({
             target: object.constructor,
             propertyName: propertyName,
             options: validationOptions,
             constraints: [],
-            validator: IsIdExistingConstraint
+            validator: IsClassIdExistingConstraint
         })
     }
 }
