+++ data/validators/IsReportIdExisting.ts
@@ -5,19 +5,18 @@ import {
     ValidatorConstraint,
     ValidatorConstraintInterface
 } from "class-validator";
-import { MysqlDatabase } from "../../../database/MysqlDatabase";
-import { UsersModel } from "../../../models/UsersModel";
+import { MysqlDatabase } from "../../database/MysqlDatabase";
+import { ReportsModel } from "../../models/ReportsModel";
 
 @ValidatorConstraint({async: true})
-export class IsUserIdExistingConstraint implements ValidatorConstraintInterface {
+export class IsReportIdExistingConstraint implements ValidatorConstraintInterface {
     validate(id: number, validationArguments?: ValidationArguments): Promise<boolean> | boolean {
         return new Promise(async resolve => {
             if (!id) {
-
                 resolve(true);
             }
 
-            const result = await new UsersModel(new MysqlDatabase())
+            const result = await new ReportsModel(new MysqlDatabase())
                 .findById(id);
 
             resolve(!!result);
@@ -25,14 +24,14 @@ export class IsUserIdExistingConstraint implements ValidatorConstraintInterface 
     }
 }
 
-export function IsUserIdExisting(validationOptions: ValidationOptions) {
+export function IsReportIdExisting(validationOptions: ValidationOptions) {
     return function (object: Object, propertyName: string) {
         registerDecorator({
             target: object.constructor,
             propertyName: propertyName,
             options: validationOptions,
             constraints: [],
-            validator: IsUserIdExistingConstraint
+            validator: IsReportIdExistingConstraint
         })
     }
 }
